выполняю задание 2-2
