+++ index.php
@@ -1,47 +1,8 @@
-<?php
-class Advert {
-    public int $rooms;
-    public string $category;
-    public int $price;
-    public string $type;
-    public string $period;
+ <?php
+ require_once 'Sale.php';
+ require_once 'Rent.php';
+ require_once 'Advert.php';
 
-        public function __construct($rooms, $category, $price, $type, $period) {
-            $this->rooms = $rooms;
-            $this->category = $category;
-            $this->price = $price;
-            $this->type = $type;
-            $this->period = $period;
-        }
-        
-        public function getTitle() {
-            $result = "";
-            if ($this->category == "sale") {
-                $result .= "Продам ";
-            } else if ($this->category == "rent") {
-                $result .= "Сдам ";
-            }
-            if ($this->type == "dom") {
-                $result .= $this->rooms." - комнатный дом";
-            } else {
-                $result .= $this->rooms." - комнатную квартиру";
-            }
-            if ($this->price > 1000000) {
-                $result .= " за ".substr($this->price, 0, 2)." млн тг";
-            } else {
-                $result .= " за ". $this->price." тг";
-            }
-            if ($this->period == "month") {
-                $result .= " в месяц";
-            } else if ($this->period == "day") {
-                $result .= " в сутки";
-            } else if ($this->period == NULL) {
-                $result .= "";
-            }
-            
-            return $result;
-  }
-}
 $adverts = [
     ['rooms' => 5, 'category' => 'sale', 'price' => 55000000, 'type' => 'dom'],
     ['rooms' => 2, 'category' => 'sale', 'price' => 21500000, 'type' => 'kvartira'],
@@ -49,7 +10,13 @@ $adverts = [
     ['rooms' => 1, 'category' => 'rent', 'price' => 15000, 'type' => 'kvartira', 'period' => 'day'],
 ];
     foreach ($adverts as $value) {
-        $adv = new Advert($value['rooms'], $value['category'], $value['price'], $value['type'], $value['period'] ?? "");
-        echo $adv->getTitle()."<br>";
+        if ($value['category'] == "sale"){
+            $adv = new Sale($value['rooms'], $value['category'], $value['price'], $value['type'], $value['period'] ?? "");
+            echo $adv->getTitle()."<br/>";
+        }else{
+            $ads = new Rent($value['rooms'], $value['category'], $value['price'], $value['type'], $value['period'] ?? "");
+            echo $ads->getTitle()."<br/>";
+        }
+        
 }
 ?>
